解决了dijkstra的bug，是因为加权有向边的构造函数里面的权重是int类型，但是读入参数去构造边的时候是double类型，类型自动转换后使得所有边的权值都是0。还有一个问题是逐个点进行松弛的时候应该先松弛起点。

+++ Algorithms/AlgorithmAndDatastructure/EdgeWeightedGraph/DirectedEdge.cpp
@@ -5,7 +5,7 @@
 #include <sstream>
 #include "DirectedEdge.h"
 
-DirectedEdge::DirectedEdge(int v, int w, int weight) {
+DirectedEdge::DirectedEdge(int v, int w, double weight) {
     this->v = v;
     this->w = w;
     this->weight = weight;
